Rework Definition data structure. We have lift-off
I found a handy tool that converts JSON objects to classes of
your language of choice. It wasn't foolproof, and required quite
a lot of patching, due to the newer Dart standard being quite
null-proof. Was able to use the Dart Migration tool to fix that.
The definition, on the debug page, now displays! Time to incorporate
a couple more of the objects from this data structure into the page.