Add API code, and add more testing code
I implemented some of the API code multiple times, in multiple places. I
aim to completely refactor the API code, quite close to what I did in
the merriam_webster_api.dart file.

Essentially, the two tasks being
performed here need to be split up into two classes. fetchFromApi and
parseJson or something like this. This way, the fetch stage is seperate
from the parse stage. We can then fetch json responses from the API
asynchronously, without needing to also parse the response
asynchronously.

The problem is, if we use a FutureBuilder, in the page that features the
definition, we can only query the definition once, in the initState()
function. After this initial query, it's difficult to get Flutter to
update the state of that variable, when using a FutureBuilder, so I've
got some work to do, implementing a custom sort of waiting logic, for if
the connection is slow, and it takes longer to fetch from the API. Or if
no connection is made, etc.