Fix test cases with test_page.dart
Located the source of the problem, when the bdd_widget_tests generate
tests, one of them will require you to modify the constructor call in
order to run the test. The original example used MyApp(), so I found the
corresponding function call in my example class, and it works!

For the third test, I was following Oleksandr Leushchenko's tutorial on
YouTube, titled "BDD in Flutter - how to test mobile apps in a natural
language"