Fix Hangman, so new word resets Definition
I made a few changes to hangman, so when you either go back to home,
or click new window, upon winning or losing, the word is displayed,
and the definition is displayed. I was able to find a way to get the
new definiton to display, when you "refresh" the page, by clicking
new page.