Ammend Hangman Test Cases and Add Debug Mode
I refactored most of the hangman test cases to line up with
a new class in hangman_debug.dart. This way we can test in
"Debug Mode", but testing the rest of the cases, with the
Page quering the API, this will cause the tests all to fail.
Flutter does NOT allow tests to connect to the internet. All
test cases where we make an API call, we will have to cache
actual API responses, and mock each test case, with Mockito.